+++ src/index.ts
@@ -3,6 +3,7 @@
 //
 
 export * from "./components/forms/Form.js";
+export * from "./components/forms/FormCheckboxInput.js";
 export * from "./components/forms/FormFieldset.js";
 export * from "./components/forms/FormNoticeContainer.js";